backend/processing/chunking_embeddings_gpu/requirements.txt
python-dotenv
sentence-transformers
qdrant-client
nltk
tqdm
torch
